move SideNavbar to components
