+++ src/components/TopNavbar/TopNavbar.style.tsx
@@ -1,10 +1,11 @@
 import styled from '@emotion/styled'
 
-import { Searchbar, SIDENAVBAR_WIDTH } from '@/shared/components'
+import { Searchbar } from '@/shared/components'
 import { breakpoints, colors, sizes, transitions, zIndex } from '@/shared/theme'
 import { ReactComponent as UnstyledShortLogo } from '@/assets/logo.svg'
 import { ReactComponent as UnstyledFullLogo } from '@/assets/full-logo.svg'
 import { Link } from '@reach/router'
+import { SIDENAVBAR_WIDTH } from '@/components/SideNavbar'
 
 type TopNavbarStyleProps = {
   hasFocus: boolean
