+++ src/components/index.ts
@@ -12,3 +12,4 @@ export { default as ErrorFallback } from './ErrorFallback'
 export { default as ChannelLink } from './ChannelLink'
 export { default as BackgroundPattern } from './BackgroundPattern'
 export { InfiniteVideoGrid, InfiniteChannelGrid } from './InfiniteGrids'
+export { default as SideNavbar, SIDENAVBAR_WIDTH, EXPANDED_SIDENAVBAR_WIDTH, NavItem } from './SideNavbar'
